assignment_4/Module_4/requirements.txt
Flask
psycopg
beautifulsoup4
urllib3
huggingface_hub
llama-cpp-python
pytest
pytest-cov
sphinx
sphinx-rtd-theme
